+++ scanpipe/tests/pipes/test_output.py
@@ -247,7 +247,7 @@ class ScanPipeOutputPipesTest(TestCase):
 
         project = Project.objects.get(name="asgiref")
         package = project.discoveredpackages.get(
-            uuid="b2d24c22-0dff-4e3f-8332-413b4f4852a7"
+            uuid="8fc54317-6054-4eb7-b189-f1b2822d3fa6"
         )
 
         package.other_license_expression_spdx = "Apache-2.0 AND LicenseRef-test"