+++ CHANGELOG.rst
@@ -1,7 +1,17 @@
 Changelog
 =========
 
-v34.6.3 (unreleased)
+v34.6.4 (unreleased)
+--------------------
+
+- Add all "classify" plugin fields from scancode-toolkit on the CodebaseResource model.
+  https://github.com/nexB/scancode.io/issues/1275
+
+- Refine the extraction errors reporting to include the resource path for rendering
+  link to the related resources in the UI.
+  https://github.com/nexB/scancode.io/issues/1273
+
+v34.6.3 (2024-06-21)
 --------------------
 
 - Use the ``--option=value`` syntax for args entries in place of ``--option value``
@@ -11,9 +21,8 @@ v34.6.3 (unreleased)
 - Fix an issue in the d2d JavaScript mapper.
   https://github.com/nexB/scancode.io/pull/1274
 
-- Refine the extraction errors reporting to include the resource path for rendering
-  link to the related resources in the UI.
-  https://github.com/nexB/scancode.io/issues/1273
+- Add support for a ``ignored_vulnerabilities`` field on the Project configuration.
+  https://github.com/nexB/scancode.io/issues/1271
 
 v34.6.2 (2024-06-18)
 --------------------